Update raman_data_generator.py
I was slicing the samples due in order to clip out some distortions of the signal generated by synthetic interpolation on my data. However it actually should be done before the data are given to the generator itself so I'm removing slicing. Doing so allows also to keep untouched the shape of the samples